pyproject.toml
[project]
name = "blogbackend2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohappyeyeballs==2.4.4",
    "aiohttp==3.11.11",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "apscheduler==3.11.0",
    "async-timeout==5.0.1",
    "attrs==25.1.0",
    "beautifulsoup4==4.12.3",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "deprecation==2.1.0",
    "dnspython==2.7.0",
    "email-validator==2.2.0",
    "exceptiongroup==1.2.2",
    "fastapi==0.115.8",
    "fastapi-cli==0.0.7",
    "feedparser==6.0.11",
    "frozenlist==1.5.0",
    "gnews==0.3.9",
    "h11==0.14.0",
    "h2==4.1.0",
    "hpack==4.1.0",
    "httpcore==1.0.7",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "hyperframe==6.1.0",
    "idna==3.10",
    "iniconfig==2.1.0",
    "jinja2==3.1.5",
    "markdown==3.6",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "multidict==6.1.0",
    "packaging==24.2",
    "pillow==11.3.0",
    "pluggy==1.6.0",
    "propcache==0.3.2",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pygments==2.19.1",
    "pymongo==4.13.2",
    "pytest==8.4.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "requests==2.32.3",
    "rich==13.9.4",
    "rich-toolkit==0.13.2",
    "sgmllib3k==1.0.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.6",
    "starlette==0.45.3",
    "typer==0.15.1",
    "typing-extensions==4.12.2",
    "tzlocal==5.2",
    "urllib3==2.3.0",
    "uvicorn==0.34.0",
    "uvloop==0.21.0",
    "watchfiles==1.0.4",
    "websockets==14.2",
    "yarl==1.18.3",
]
